requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
pyyaml>=6.0.1
pydantic>=2.5.0
python-dotenv>=1.0.0
tenacity>=8.2.0
aiosqlite>=0.19.0
pyjwt>=2.8.0
